component relocation and additions
